🚚 change parallel route to main route
