+++ app/(home)/@rulecards/page.tsx
@@ -1,12 +1,10 @@
 import RuleCard from "@components/home/client/rule-card";
 import { reader } from "app/reader";
-import { Suspense, useMemo } from "react";
+import { Suspense } from "react";
 
 const RulesCardPage = async () => {
 	const allRules = await reader.collections.rules.all()
 	const categories = await reader.collections.category.all()
-
-
 
 	return (
 		<Suspense fallback={<>Loading...</>}>
