+++ app/(home)/@trending/default.tsx
